Create Pattern Matching
Pattern Matching
Now that you know how to iterate through a 2D array, your next task will be to use what you've learned to help locate a robot in a 2D world based on some sensor measurements.

In this example, a robot is navigating a 4x5 orange and blue world, shown below. This robot also has a sensor that tells it the color of the square it's on top of and the color of the square directly to its right.


2D orange and blue world grid


Robot sensing two color squares (one it's on top of, and one to its right)

In Python code, the world and the sensor measurement can be represented by the following code (where 'o' stands for orange and 'b' for blue):

import numpy as np

world = np.array ([ ['o', 'b', 'o', 'o', 'b'],
                    ['o', 'o', 'b', 'o', 'o'],
                    ['b', 'o', 'o', 'b', 'o'],
                    ['b', 'o', 'o', 'o', 'o'] ])

measurement = ['b', 'o']
It's now easy for us to see that there are only a few places in the world that this robot can be, based on where the measurement lines up with color patterns on the world grid! In this exercise, it'll be up to you to find these locations programmatically.

Hint: You'll be adding values to a list. To add indices, please use the append function and the format list_name.append([row,column]). Further instructions can be found in the TODOs below.

pattern_matching.py